Bug fix on set selectors

+++ dbmodel/utils/fct/gw_fct_setselectors.sql
@@ -116,7 +116,7 @@ BEGIN
 		IF v_value THEN
 			EXECUTE 'INSERT INTO ' || v_tablename || ' ('|| v_columnname ||', cur_user) VALUES('|| v_id ||', '''|| current_user ||''')ON CONFLICT DO NOTHING';
 		ELSE
-			EXECUTE 'DELETE FROM ' || v_tablename || ' WHERE ' || v_columnname || ' = '|| v_id ||'';
+			EXECUTE 'DELETE FROM ' || v_tablename || ' WHERE ' || v_columnname || ' = '|| v_id ||' AND cur_user = current_user';
 		END IF;
 
 	END IF;
